Split order_cycle_spec into smaller specs grouped by theme
